requirements.txt
streamlit
groq
faiss-cpu
sentence-transformers
requests